fjerner interfaces for impls som bare kalles fra en jobb
