--- service/src/main/kotlin/no/nav/su/se/bakover/service/skatt/JournalførSkattDokumentServiceImpl.kt
@@ -1,59 +1,0 @@
-package no.nav.su.se.bakover.service.skatt
-
-import arrow.core.Either
-import arrow.core.getOrElse
-import arrow.core.left
-import no.nav.su.se.bakover.client.dokarkiv.DokArkiv
-import no.nav.su.se.bakover.client.dokarkiv.Journalpost
-import no.nav.su.se.bakover.client.dokarkiv.JournalpostSkatt.Companion.lagJournalpost
-import no.nav.su.se.bakover.common.journal.JournalpostId
-import no.nav.su.se.bakover.domain.brev.KunneIkkeJournalføreDokument
-import no.nav.su.se.bakover.domain.person.PersonService
-import no.nav.su.se.bakover.domain.sak.SakService
-import no.nav.su.se.bakover.domain.skatt.DokumentSkattRepo
-import no.nav.su.se.bakover.domain.skatt.Skattedokument
-import no.nav.su.se.bakover.service.dokument.DokumentResultatSet
-import no.nav.su.se.bakover.service.dokument.DokumentResultatSet.Companion.logResultat
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-
-class JournalførSkattDokumentServiceImpl(
-    private val dokArkiv: DokArkiv,
-    private val sakService: SakService,
-    private val personService: PersonService,
-    private val dokumentSkattRepo: DokumentSkattRepo,
-) : JournalførSkattDokumentService {
-    val log: Logger = LoggerFactory.getLogger(this::class.java)
-
-    override fun journalfør() {
-        val skatteDokumenterSomMåJournalføres = dokumentSkattRepo.hentDokumenterForJournalføring()
-
-        skatteDokumenterSomMåJournalføres.map { skattedokument ->
-            journalførSkattedokument(skattedokument)
-                .map { DokumentResultatSet.Ok(it.id) }
-                .mapLeft { DokumentResultatSet.Feil(skattedokument.id) }
-        }.logResultat(log)
-    }
-
-    private fun journalførSkattedokument(skattedokument: Skattedokument.Generert): Either<KunneIkkeJournalføreDokument, Skattedokument.Journalført> {
-        val sakInfo = sakService.hentSakInfo(skattedokument.sakid).getOrElse {
-            throw IllegalStateException("Fant ikke sak. Her burde vi egentlig sak finnes. sakId ${skattedokument.sakid}")
-        }
-        val person = personService.hentPersonMedSystembruker(sakInfo.fnr)
-            .getOrElse { return KunneIkkeJournalføreDokument.KunneIkkeFinnePerson.left() }
-
-        return journalfør(skattedokument.lagJournalpost(person, sakInfo)).map {
-            val tilJournalført = skattedokument.tilJournalført(it)
-            dokumentSkattRepo.lagre(tilJournalført)
-            tilJournalført
-        }
-    }
-
-    private fun journalfør(journalpost: Journalpost): Either<KunneIkkeJournalføreDokument, JournalpostId> {
-        return dokArkiv.opprettJournalpost(journalpost)
-            .mapLeft {
-                log.error("Journalføring: Kunne ikke journalføre i eksternt system (joark/dokarkiv)")
-                KunneIkkeJournalføreDokument.FeilVedOpprettelseAvJournalpost
-            }
-    }
-}
